requirements.txt
boto3
python-dotenv
snowflake-connector-python
pandas
